feat: Implement versioned updates for tickets and inquiries
- Refactored ticket and inquiry update methods to include expected versioning, allowing for conflict detection and resolution.
- Introduced CaseMutationResult and CaseMutationStatus enums to handle update outcomes more effectively.
- Updated API endpoints to return appropriate responses based on mutation status.
- Enhanced SignalR event publishing to include new ticket and inquiry states.
- Modified front-end JavaScript to support versioned updates for ticket and inquiry status changes.
- Added a new case audit table to track changes and maintain an append-only history for ticket and inquiry mutations.
- Updated data models and contracts to include versioning information for tickets and inquiries.

+++ CBSSupport.Shared/Services/ConversationOutboxRepository.cs
@@ -54,7 +54,7 @@ public sealed class ConversationOutboxRepository(
                    message.id AS MessageRecordId,
                    message.instruction AS MessageText,
                    message.datetime AS MessageSentAt,
-                   COALESCE(message.insert_user, message.client_auth_user_id) AS SenderUserId,
+                   COALESCE(message.insert_user, message.client_auth_user_id)::bigint AS SenderUserId,
                    CASE WHEN message.id IS NULL THEN NULL
                         WHEN message.client_auth_user_id IS NULL THEN 'Admin'
                         ELSE 'Client' END AS SenderKind,
